==> Lessons/02_Lab_Intro_DataFrame_lj.Rmd <==
---
title: "Intro to Data Frames in R"
author: "Environmental Data Analytics | John Fay and Luana Lima"
date: "Fall 2022"
output: pdf_document
editor_options: 
  chunk_output_type: inline
---
## setting initial code. want all R code chunks to be in the report when the final report is knitted
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Data frame definition

A **data frame** is used for storing data tables. It a list a vector of equal length. 

When we import data to R, data frame is the preferred way for storing the data because columns can have different modes (character, numeric, integer, logical, complex).


## Data frame built-in example

Let's look into a built-in data frame from package "datasets" - cars. The data give the speed of cars and the distances taken to stop.

```{r cars}
cars
```


Note that it has 2 columns and 50 rows.

## Data frame columns

Suppose you want just the column speed. How would you access that data? 
# $ refers to column inside a dataframe
```{r}
cars$speed

```

How would you store it on another object?
# car_speed is name of object where you want the data to be stored
```{r}
car_speed <- cars$speed
car_speed
```
# [1] or [32] value number it starts with 

## Transforming object in a data frame

Is the new object you create a data frame?
```{r}
#Option 1
class(car_speed)

#Option 2
is.data.frame(car_speed)
```

How could you make it a data frame?
# as.data.frame function transform vectors into dataframe
```{r}
df_car_speed <- as.data.frame(car_speed)
df_car_speed
class(df_car_speed)
```

## Adding columns to a data frame

How could you add columns to *df_car_speed*?

```{r}
car_dist <- cars$dist

#Option 1
# cbind stands for column bind - will bind 2 objects by column - cbind wont work if there are different number of rows
df <- cbind(df_car_speed,car_dist)  #similarly rows could be added using rbind()
class(df)
df

#Or Option 2 - transform into a data frame before binding
df_car_dist <- as.data.frame(car_dist)  #op2
df_opt2 <- cbind(df_car_speed,df_car_dist)
class(df_opt2)
df_opt2

```

Note that when we transformed the vector in a data frame the name of the vector became the column name.

```{r}
colnames(df)  #or simply names()
names(df)
```

## Creating a data frame

How would you create a data frame?

```{r}
#useful function data.frame() -> arguments these functions will take is the columns which u want it to be displayed as
create_df <- data.frame("speed"=car_speed,"dist"=car_dist)
create_df #this line helps to print out what the code does, so we can visualise
```


## Data frame functions

Some useful functions to use with data frames.

```{r}
ncol(df)  #number of columns in the dataframe
nrow(df)  #number of rows in the dataframe
length(df)  #same as ncol, n is with respect to the number of variables in the data frame 
summary(df) #summary statistics of the data 
head(df) #show the first 6 rows of df - the head of the dataframe 

#If you know the number of the column you want you can refer to that to access column
df_2 <- df[1:10,1:2]

#you could also use this notation to delete columns
df <- df[,-2]
df
```
