==> Lessons/06_Lab_GLMs_lj.Rmd <==
---
title: "6: Part 1 - Generalized Linear Models"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Fall 2022"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
1. Answer questions on M5/A5
2. Answer questions on M6 - GLMs
3. Practice more application GLM to real datasets


## Set up
```{r, message = FALSE}
library(tidyverse)
library(agricolae)   

EPAair <- read.csv("./Data/Processed/EPAair_O3_PM25_NC1819_Processed.csv", stringsAsFactors = TRUE)
# Set date to date format
EPAair$Date <- as.Date(EPAair$Date, format = "%Y-%m-%d")

Litter <- read.csv("./Data/Processed/NEON_NIWO_Litter_mass_trap_Processed.csv", stringsAsFactors = TRUE)
# Set date to date format
Litter$collectDate <- as.Date(Litter$collectDate , format = "%Y-%m-%d")

# Set theme
mytheme <- theme_classic(base_size = 14) +
  theme(axis.text = element_text(color = "black"), 
        legend.position = "top")
theme_set(mytheme)
```



## Visualization and interpretation challenge

Create three plots, each with appropriately formatted axes and legends. Choose a non-default color palette.

1. geom_density of ozone divided by year (distinguish between years by adding transparency to the geom_density layer).
2. geom_boxplot of ozone divided by year. Add letters representing a significant difference between 2018 and 2019 (hint: stat_summary). 
3. geom_violin of ozone divided by year, with the 0.5 quantile marked as a horizontal line. Add letters representing a significant difference between 2018 and 2019. 

```{r}
#Exercise 1: 

O3.denplot <-
  ggplot(EPAair, aes(x=Ozone, fill=as.factor(Year))) +
  geom_density() +
  scale_fill_manual(values=c("darkgreen", "darkblue")) + 
  labs (x="Ozone AQI value", y="Density", fill="")

O3.denplot


#Exercise 2

O3.boxplot <-
  ggplot2::ggplot(EPAair, aes(x=as.factor(Year), y=Ozone)) +
  geom_boxplot() +
  stat_summary(geom = "text", fun=max, vjust= -1, size=4,
               label=c("b", "a")) +
  labs(x="", y="Ozone AQI value") +
  ylim(0,150)

O3.boxplot


#when you look at mean, at first quantile, they are v close to each other. max and mean value shows the greatest difference
  
  # Ex 3
  
 O3.violin <-
  ggplot(EPAair, aes(x=as.factor(Year), y=Ozone)) +
  geom_violin(draw_quantiles = 0.5) +
  stat_summary(geom = "text", fun=max, vjust= -1, size=4,
               label=c("b", "a")) +
  labs(x="", y="Ozone AQI value") +
  ylim(0,150)

O3.violin

```

## Linear Regression

Important components of the linear regression are the correlation and the R-squared value. The **correlation** is a number between -1 and 1, describing the relationship between the variables. Correlations close to -1 represent strong negative correlations, correlations close to zero represent weak correlations, and correlations close to 1 represent strong positive correlations. The **R-squared value** is the correlation squared, becoming a number between 0 and 1. The R-squared value describes the percent of variance accounted for by the explanatory variables. 

For the NTL-LTER dataset, can we predict PM2.5 from Ozone? #predicting pm2.5, thats why its y, then ozone is x.

```{r}

#Exercise 2: Run a linear regression PM2.5 by Ozone. Find the p-value and R-squared value. 

LinearRegression <- lm(data=EPAair, PM2.5~ Ozone)
summary(LinearRegression)


#statistically signinficant, can use ozone to predict pm2.5

irradiance.regression <- lm(data = PeterPaul.chem.nutrients, irradianceWater ~ depth)
summary(irradiance.regression)
#Exercise 3: Build a scatterplot. Add a line and standard error for the linear regression. Add the regression equation to the plot

PM2.5byOzone.Plot <-
  ggplot(EPAair, aes(x=Ozone, y=PM2.5)) +
  geom_point() +
  geom_smooth(method = "lm", col="blue")
print(PM2.5byOzone.Plot)


```


## AIC to select variables

What other variables can we add to improve model?

```{r}

#Exercise 4: Build correlation plots and identify more possible explanatory variables to add to the regression.

library(corrplot)

#model answer. as long as numeric, can include them as explanatory variables
EPAAir.variables2 <-
  EPAair %>%
  select(meanLAT:Ozone) %>%
  na.omit()

EPAairCorr2 <- cor(EPAAir.variables2) #create a matrix that has correlation between each one of the columns with each other. symmetric matrix. 
#these are more visual representations of the matrix. orange to reds are negative correlations, blues are positive correlations
corrplot(EPAairCorr2, method ="ellipse")
corrplot.mixed(EPAairCorr2, upper="ellipse")

#Exercise 5: Choose a model by AIC in a Stepwise Algorithm. Do the results from AIC match the variables you selected on Exercise 5?


# running linear model with all the variables
ozonebyall.regression <- lm(data=EPAAir.variables2, PM2.5 ~ Ozone + Year + Month + meanLONG + meanLAT)
summary(ozonebyall.regression)
#rsquare is 0.192 -> explaining variability than pm2.5 that just ozone. 
#p values can tell us assumptions. p value for ozone is v small, so highly correlated with pm2.5. year, greater than 0.05, reject null, no significant correlation. month has low p value, significant correlations with pm2.5. similar to corrplot

step(ozonebyall.regression)


#my answer
EPAairAIC <- lm(data=EPAAir.variables2, PM2.5 ~ meanLAT + meanLONG +Year + Month + Ozone )

step(EPAairAIC)

summary(EPAairAIC)


#Exercise 6: Run another regression using the variables selected on Exercise 6. Compare r-squared value with the one from Exercise 5. 

Ozonebestregression <- lm(data=EPAAir.variables2, PM2.5 ~ meanLAT + meanLONG+ Month + Ozone)
summary (Ozonebestregression)

#all pvalue <0.05, r2: 0.191 -> better than earlier 


```



## Litter Exercise

```{r}
# Wrangle the data
Litter.Totals <- Litter %>%
  group_by(plotID, collectDate, nlcdClass) %>%
  summarise(dryMass = sum(dryMass))

# Format ANOVA as aov
Litter.Totals.anova <- aov(data = Litter.Totals, dryMass ~ plotID)
summary(Litter.Totals.anova)

# Extract groupings for pairwise relationships
Litter.Totals.groups <- HSD.test(Litter.Totals.anova, "plotID", group = TRUE)
Litter.Totals.groups$groups

Litter.Totals <- Litter %>%
  mutate( treatgroups = Litter.Totals.groups$groups[plotID,2])

# Graph the results
Litter.Totals.plot <- ggplot(Litter.Totals, aes(x = plotID, y = dryMass)) +
  geom_boxplot() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  stat_summary(geom = "text", fun = max, vjust = -1, size = 3.5,
               label = c("abc", "ab", "abcd", "bcd", "bcd", "a", 
                         "d", "abcd", "cd", "d", "abcd", "abcd")) +
  labs(x = "Plot ID", y = "Dry Mass (g)") +
  ylim(0, 85)
print(Litter.Totals.plot)

```

```{r}
#Exercise 7: Improve the plot to print all the As together, all the bs together etc 

#arrange by treatgroups




```
