==> Lessons/06_Part2_GLMs_ANOVA_lj.Rmd <==
---
title: "6: Part 2 - Generalized Linear Models"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Fall 2022"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
1. Apply special cases of the GLM (ANOVA) to real datasets
2. Interpret and report the results of ANOVA in publication-style formats

one-way anova -> when you have one explanatory categorical variable with 2 diff levels
-need to determine whether 3 or more populations have equal means. 
-h0 is always -> means are equal. h1 is always -> not all means are equal
-assumption: follows normal distribution, population cariances are equal, observations are independent 


two-way -> 2 explanatory variables



## Set up
```{r, message = FALSE}
getwd()
library(tidyverse)
install.packages("agricolae")
library(agricolae)   #you may need to install/update package "htmltools" before installing and loading agricolae

Litter <- read.csv("./Data/Processed/NEON_NIWO_Litter_mass_trap_Processed.csv", stringsAsFactors = TRUE)

# Set date to date format
Litter$collectDate <- as.Date(Litter$collectDate , format = "%Y-%m-%d")

# Set theme
mytheme <- theme_classic(base_size = 14) +
  theme(axis.text = element_text(color = "black"), 
        legend.position = "top")
theme_set(mytheme)
```

### One-way ANOVA
A one-way ANOVA is used when you have one categorical explanatory variable with more than two categories. In R, we can run the model with the function `lm` or `aov`, the latter of which will allow us to run post-hoc tests to determine pairwise differences. 
Reject null hypothesis when at least one pair of means are not equal. test does not tell you which of the 2 means are not equal, so we need to run post-hoc tests

For example: Does total litter biomass differ among sites at the Niwot Ridge LTER? check if mean of litter biomass differs from different sites
```{r}
# Wrangle the data
Litter.Totals <- Litter %>%
  group_by(plotID, collectDate, nlcdClass) %>% 
  summarise(dryMass = sum(dryMass)) 

summary(Litter.Totals)
#check all value of plotID. -> overall mean has a high variance when compared with the minimum and maximum values. due to site location
summary(Litter.Totals$plotID) # not a balanced experiment as we have some sites with more samples than others. but still can run the Anova test

# Test for normality. 
# Note: the important assumption for GLMs is normality of residuals, 
# not necessarily normality of raw data. See below when we plot the model.
# shapiro.test() performs the Shapiro-Wilk test of the null that data follows a normal distribution (for each of the sites below) -> only for one object, filter by sites

shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_040"]) #p value less than 0.05 means reject null hypothesis. Site 040 does not follow normal dist
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_041"])
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_046"])
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_047"])
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_051"])
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_057"]) #dont reject
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_058"])
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_061"])
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_062"])
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_063"])
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_064"])
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_067"])

#results: reject null in all except two: NIWO_057 and NIWO_046
#but method is robust from small deviations from normal distribution
qqnorm(Litter.Totals$dryMass); qqline(Litter.Totals$dryMass) #data does not follow normal distributions


# Test for equal variance
# ANOVA is robust against departures from equal variance.
# bartlett.test() performs Bartlett's test of the null that the variances in each of the groups (samples) are the same. 
bartlett.test(Litter.Totals$dryMass ~ Litter.Totals$plotID)

#results: reject null i.e. variances are not equal

# Format ANOVA as aov
Litter.Totals.anova <- aov(data = Litter.Totals, dryMass ~ plotID) #dryMass: continuous/dependent variable, plotID is categorical variable
summary(Litter.Totals.anova)

#results: p<0.05 so reject null hypothesis i.e. difference between a pair of group means is statistically significant (means are not all the same)
#sumsq is difference between overall mean and each one of the values and there is diff between sum square between groups and within groups. 
#mean sq is sum sq divide by df
#f value is mean sq of plot id/mean sq of residuals. 


# Format ANOVA as lm
Litter.Totals.anova2 <- lm(data = Litter.Totals, dryMass ~ plotID)
summary(Litter.Totals.anova2) #summary from linear will have one row for each level of categorical value (so for each site). 040 represented by intercept term.

# Checking model fit and assumptions
# ANOVA is robust against departures from normality.
plot(Litter.Totals.anova2)

# Post-hoc test -> cos ANOVA results do not identify which particular differences between pairs of means are significant, so we use post hoc tests to explore the differences between multiple group means
# TukeyHSD() computes Tukey Honest Significant Differences -> is the most common for comparing all possible group pairings
TukeyHSD(Litter.Totals.anova) #outputs for each possible combination, the difference of the mean, with lower and upper end point of the confidence interval and p value for each combination
#41 and 40 are close. 


# Extract groupings for pairwise relationships
Litter.Totals.groups <- HSD.test(Litter.Totals.anova, "plotID", group = TRUE) #HSD.test(anova item, category, group=True so function will group levels have the same mean)
Litter.Totals.groups
#treatments with same letter are not significantly different, ie they all have the same mean

# Graph the results. site 40=abc, site 41=ab, site 46=abcd -> follow the order of the sites
Litter.Totals.plot <- ggplot(Litter.Totals, aes(x = plotID, y = dryMass)) +
  geom_boxplot() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  stat_summary(geom = "text", fun = max, vjust = -1, size = 3.5,
               label = c("abc", "ab", "abcd", "bcd", "bcd", "a", 
                         "d", "abcd", "cd", "d", "abcd", "abcd")) +
  labs(x = "Plot ID", y = "Dry Mass (g)") +
  ylim(0, 85)
print(Litter.Totals.plot) #all sites with a -> means not far from each other. 
```

What information does the output give us? How might we report this information in a report?

> reorder the sites, same order as group lists. a, b, c, d

## TWO-WAY ANOVA

### Main effects
A two-way ANOVA allows us to examine the effects of two categorical explanatory variables on a continuous response variable. 

For example: What if we wanted to know how litter masses varied across functional groups and land cover class?
```{r}
# Format as aov
Litter.anova.2way <- aov(data = Litter, dryMass ~ functionalGroup + nlcdClass) 
summary(Litter.anova.2way)
#df is functional group -1 thats why 7 for functionalgroup. 
#df of residuals associated with number of observations we have


# Format as lm
Litter.anova.2way2 <- lm(data = Litter, dryMass ~ functionalGroup + nlcdClass)
summary(Litter.anova.2way2)

# Run a post-hoc test for pairwise differences
TukeyHSD(Litter.anova.2way)

```

### Interaction effects

We may expect the effects of functional group and land cover to be dependent on each other. For instance, we may expect needle biomass to be larger in evergreen forests than in other land types. In this case, we might choose to run an interaction effects two-way ANOVA, which will examine the individual effects of the explanatory variables as well as the interaction of the explanatory variables.

The output gives test statistics for each explanatory variable as well as the interaction effect of the explanatory variables. If the p-value for the interaction effect is less than 0.05, then we would consider the interaction among the explanatory variables to be significant.

```{r}
# Format as aov. interaction between functionalgroup + land cover
Litter.anova.2way3 <- aov(data = Litter, dryMass ~ functionalGroup * nlcdClass) 
summary(Litter.anova.2way3)
#interaction is significant among the variables, then we look at pairwise interaction. if interaction is not significant, we interpret difference for main effects only.

# Format as lm
Litter.anova.2way4 <- lm(data = Litter, dryMass ~ functionalGroup * nlcdClass)
summary(Litter.anova.2way4)

```

If the interaction is significant, we interpret pairwise differences for the interaction. If the interaction is not significant, we interpret differences for the main effects only.

```{r}
# Run a post-hoc test for pairwise differences
TukeyHSD(Litter.anova.2way3)

Litter.interaction <- with(Litter, interaction(functionalGroup, nlcdClass)) #create new object with 24 levels (result of functionalgroup + class)

#Litter.test <- cbind(Litter,Litter.interaction)

#compare means of each one of the litter interaction levels. one way, but is result of interaction of 2 categorical variables
Litter.anova.2way5 <- aov(data = Litter, dryMass ~ Litter.interaction)

Litter.groups <- HSD.test(Litter.anova.2way5, "Litter.interaction", group = TRUE)
Litter.groups
#needles and evergreen forest cannot be compared with any other interaction here. 
#needles and grassland can be compared with needles and shrubscrub

# Plot the results
# Edit this graph to make it more visually appealing.
# Add letters to represent statistical groupings.
Litter.anova.plot <- ggplot(Litter, aes(y = dryMass, x = nlcdClass, color = functionalGroup)) +
  geom_boxplot()
print(Litter.anova.plot)

```
